macos: macOS GUI application fundamental with Go
I tried XCode but the garbage it produces are not tolerable.
And Swift syntax, especially implicit context-dependent things are
awful.

I also tried Flutter, but it requires Cocoapod, which is Ruby project
that requires global install or ASDF setup that requires Bundler (Ruby
project?). Requiring this amount of globally installed tools is a sign
of bad quality and/or DX. Glancing Flutter docs does not give me a good
impression on cross-platform development.

Go + DarwinKit seems to be the best solution regarding code cleanness
and the number of features. It uses relatively legacy(?) API but built
app run okay and there is no critical issues I faced for now.

Since the amount of work needed is unclear due to this is my first time
writing macOS app, I commit this "application fundamentals" state to the
repository.