web_builder: Setup Storybook
Manually testing styles became difficult due to Vite full-reloads the
app and the app requires a file and an action to navigates to the
target scene. I could write simple a testbed, but Storybook is widely
known and has rigid UI, even though degration of UX and feature bloat is
becoming serious issues these days.