Setup Storybook for player application
I thrown away global-per-story approach, as it comes with delay between
the story loads and globals (styles) are applied. Even it's subtle and
will be diminished on subsequential accesses due to cache, this is not
acceptable.

By separating Storybook, CI time will be faster and no hack for
Storybook's sandbox limitation.